Revert "design patter strategie"
This reverts commit 1c4807115ebbae2629fd352404edfdfb5dd5663b.

+++ include/Hexagone.h
@@ -1,8 +1,8 @@
 #ifndef HEXAGONE_H
 #define HEXAGONE_H
 
-#include <cstddef>
-using TuileId = std::size_t;
+#include <cstdint>
+using TuileId = std::uint32_t;
 
 #include <iostream>
 #include <utility>
